Clean up config module on the backend
